- rework of behavior, having some visible result with shop enter system.

+++ nogameCode/nogame/characters/car3/GenerateCharacterOperator.cs
@@ -243,7 +243,7 @@ class GenerateCharacterOperator : engine.world.IFragmentOperator
                         Speed = (30f + rnd.GetFloat() * 20f + (float)carIdx * 20f) / 3.6f
                     })
                 {
-                    MaxDistance = propMaxDistance
+                    MaxDistance = (short)propMaxDistance
                 });
 
 #if DEBUG
